go back to using Helland's kelvin->rgb algorithm
Not sure if I checked in the original version of this, but move
back to using this algorithm, so that we cover the range > 7000.

The values are very slightly different from the table that was
extracted from the govee app, but I think they are close enough
that it should be fine.